refactor: add transaction request transformer
Instead of using a custom transformation function to simplify the date
when adding a transaction, we are now using a request transformation
that takes care of the date transformation directly.

Also had to change the configuration system. node-config is looking for
the config directory not where the config.get() call takes places, but
where the packages is eventually used (for example, if config.get() is
called in core, but that particular core function is used in api, then
the config directory is also expected in api). I am now using a
simplified version of the same config logic, but the config file only
needs to be defined in one place in core.