feat: ensure transactions are unique
Added a hash column to the transactions table which stores a hash
calculated from the other columns prior to insertion. This hash is set
to be unique, and therefore duplicate transactions cannot be added to
the database.

Also changed the transaction DTO to not use store a Date object any
longer. Instead, the ISO 8601 string is generated on the client side and
transferred in the DTO.